requirements.txt
pandas
python-docx
tabulate
PySide6
docxcompose